Fix "BUG: output plugins MUST implement this method" when used with Fluentd 0.14.x
When overriding Fluent::Compat::BufferedOutput#format_stream,
Fluentd 0.14 expects the plugin class itself, but no its included module,
to define format_stream method.
format method, which is not implemented by out_norikra, is otherwise invoked and
a NotImplementedError is raised.

In order to fix this issue, this patch defines NorikraOutput#format_stream, which
just delegates to the OutputMixin module.